Removed jenkins sudo requirement
This removes the need for the jenkins user to have passwordless sudo
by downloading the slave.jar file as the current user and setting the
user and group to the jenkins user.

This also replaces the cp shell command with a synchronize module, which
means that task will only be marked as changed when the files actually
change.